Refactored quiz query
